// packages/jsonpath/lib/query/syntax-tree-type.ts
/**
 * Type of a syntax tree.
 */
export enum SyntaxTreeType {
    dollarToken = "DollarToken",
    atToken = "AtToken",
    starToken = "StarToken",
    questionMarkToken = "QuestionMarkToken",
    dotToken = "DotToken",
    doubleDotToken = "DoubleDotToken",
    commaToken = "CommaToken",
    colonToken = "ColonToken",
    doubleAmpersandToken = "DoubleAmpersandToken",
    doubleBarToken = "DoubleBarToken",
    exclamationMarkToken = "ExclamationMarkToken",
    doubleEqualsToken = "DoubleEqualsToken",
    exclamationMarkEqualsToken = "ExclamationMarkEqualsToken",
    lessThanEqualsToken = "LessThanEqualsToken",
    greaterThanEqualsToken = "GreaterThanEqualsToken",
    lessThanToken = "LessThanToken",
    greaterThanToken = "GreaterThanToken",
    trueToken = "TrueToken",
    falseToken = "FalseToken",
    nullToken = "NullToken",
    stringToken = "StringToken",
    numberToken = "NumberToken",
    nameToken = "NameToken",
    openingParanthesisToken = "OpeningParanthesisToken",
    closingParanthesisToken = "ClosingParanthesisToken",
    openingBracketToken = "OpeningBracketToken",
    closingBracketToken = "ClosingBracketToken",
    endOfFileToken = "EndOfFileToken",
    missingToken = "MissingToken",
    query = "Query",
    subQuery = "SubQuery",
    segment = "Segment",
    nameSelector = "NameSelector",
    wildcardSelector = "WildcardSelector",
    sliceSelector = "SliceSelector",
    indexSelector = "IndexSelector",
    filterSelector = "FilterSelector",
    missingSelector = "MissingSelector",
    orExpression = "OrExpression",
    andExpression = "AndExpression",
    notExpression = "NotExpression",
    paranthesisExpression = "ParanthesisExpression",
    comparisonExpression = "ComparisonExpression",
    filterQueryExpression = "FilterQueryExpression",
    functionExpression = "FunctionExpression",
    numberLiteralExpression = "NumberLiteralExpression",
    stringLiteralExpression = "StringLiteralExpression",
    booleanLiteralExpression = "BooleanLiteralExpression",
    nullLiteralExpression = "NullLiteralExpression",
    missingExpression = "MissingExpression"
}